Added Collision Sound and Level Clear
Two objects have collision that changes based on speed. The speed should be relative to object weight but I am uncertain if I got that right.

Level clear is a reset at the moment, it can be changed to allow for transition to a new level later.

The patch should now begin with the main menu first.